Switch to multi-bucket tt. No significant elo change expected.
Result:
Score of Loki DEV vs Loki DEV baseline: 727 - 671 - 1098  [0.511] 2496
...      Loki DEV playing White: 475 - 213 - 560  [0.605] 1248
...      Loki DEV playing Black: 252 - 458 - 538  [0.417] 1248
...      White vs Black: 933 - 465 - 1098  [0.594] 2496
Elo difference: 7.8 +/- 10.2, LOS: 93.3 %, DrawRatio: 44.0 %
SPRT: llr 1.03 (35.1%), lbound -2.94, ubound 2.94

+++ Loki.Lib/loki.pch.hpp
@@ -598,7 +598,7 @@ namespace loki::search
 		using tri_pv_table_t = std::unique_ptr<tri_pv_table<>>;
 	}
 
-	enum ttFlag : uint16_t
+	enum ttFlag : uint8_t
 	{
 		FLAG_ALPHA = 0,
 		FLAG_BETA = 1,
